refactor(#12): 이메일 인증 단계 컴포넌트명 명확화
