+++ src/components/EmailVerification/VerificationCodeStep/VerificationCodeStep.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import PinCodeInput from '../../_common/molecules/PinCodeInput/PinCodeInput';
-import S from './EmailVerificationStep2.styles';
+import S from './VerificationCodeStep.styles';
 import Button from '../../_common/atoms/Button/Button';
 import LinkedText from '../../_common/atoms/LinkedText/LinkedText';
 
-interface EmailVerificationStep2Props {
+interface VerificationCodeStepProps {
   email: string;
   nextStep: () => void;
 }
 
-export default function EmailVerificationStep2({ email, nextStep }: EmailVerificationStep2Props) {
+export default function VerificationCodeStep({ email, nextStep }: VerificationCodeStepProps) {
   const handleStep2Submit = () => {
     // todo: 이메일 인증번호 검증 api
     console.log('Step 2 complete');
@@ -18,7 +18,7 @@ export default function EmailVerificationStep2({ email, nextStep }: EmailVerific
   const handleResend = () => {
     // todo: 이메일 재전송 api
     console.log('Resend email');
-  }
+  };
 
   const handlePress = (func: any) => {
     func();
@@ -34,11 +34,23 @@ export default function EmailVerificationStep2({ email, nextStep }: EmailVerific
           <S.Label>{email}으로 전송된 인증 코드 6자리를 입력해 주세요.</S.Label>
           <PinCodeInput pinLength={6} />
           <S.InfoWrapper>
-            <LinkedText onPress={() => {handlePress(handleResend)}}>인증 코드가 오지 않나요?</LinkedText>
+            <LinkedText
+              onPress={() => {
+                handlePress(handleResend);
+              }}
+            >
+              인증 코드가 오지 않나요?
+            </LinkedText>
             <S.Timer>5:00</S.Timer>
           </S.InfoWrapper>
         </S.VerificationContainer>
-        <Button onPress={() => {handlePress(handleStep2Submit)}}>다음</Button>
+        <Button
+          onPress={() => {
+            handlePress(handleStep2Submit);
+          }}
+        >
+          다음
+        </Button>
       </S.Layout>
     </S.AppContainer>
   );
